refactor ZoneView form
